+++ ChangeLog.md
@@ -1,5 +1,10 @@
 2020-02
 
+- db schema: mission template now with a `retrieved_at` column to indicate when
+  the info was last retrieved from the provider.
+  At insertion time, this will be `null` for a directory, and a datetime for
+  a particular template.
+
 - db schema: mission template relations only to actual templates, not directories
 
 - adding handling for directory-based dispatch of mission templates